64-bit compile fixes

+++ renderdoc/driver/gl/wrappers/gl_texture_funcs.cpp
@@ -3216,7 +3216,7 @@ bool WrappedOpenGL::Serialise_glTextureBufferRangeEXT(GLuint texture, GLenum tar
 	if(m_State == READING)
 	{
 		ResourceId liveId = GetResourceManager()->GetLiveID(texid);
-		m_Textures[liveId].width = uint32_t(Size)/GetByteSize(1, 1, 1, GetBaseFormat(fmt), GetDataType(fmt), 1);
+		m_Textures[liveId].width = uint32_t(Size)/uint32_t(GetByteSize(1, 1, 1, GetBaseFormat(fmt), GetDataType(fmt), 1));
 		m_Textures[liveId].height = 1;
 		m_Textures[liveId].depth = 1;
 		m_Textures[liveId].curType = TextureTarget(Target);
@@ -3280,7 +3280,7 @@ bool WrappedOpenGL::Serialise_glTextureBufferEXT(GLuint texture, GLenum target, 
 		ResourceId liveId = GetResourceManager()->GetLiveID(texid);
 		uint32_t Size = 1;
 		m_Real.glGetNamedBufferParameterivEXT(buffer, eGL_BUFFER_SIZE, (GLint *)&Size);
-		m_Textures[liveId].width = Size/GetByteSize(1, 1, 1, GetBaseFormat(fmt), GetDataType(fmt), 1);
+		m_Textures[liveId].width = Size/uint32_t(GetByteSize(1, 1, 1, GetBaseFormat(fmt), GetDataType(fmt), 1));
 		m_Textures[liveId].height = 1;
 		m_Textures[liveId].depth = 1;
 		m_Textures[liveId].curType = TextureTarget(Target);
